+++ torchao/dtypes/__init__.py
@@ -1,13 +1,13 @@
 from . import affine_quantized_tensor_ops
 from .affine_quantized_tensor import (
     AffineQuantizedTensor,
-    to_affine_quantized_float8,
     to_affine_quantized_floatx_static,
     # experimental, will be merged into floatx in the future
     to_affine_quantized_fpx,
     to_affine_quantized_intx,
     to_affine_quantized_intx_static,
 )
+from .float8 import to_affine_quantized_float8
 from .floatx import Float8Layout
 from .nf4tensor import NF4Tensor, to_nf4
 from .uintx import (
@@ -19,8 +19,8 @@ from .uintx import (
     MarlinSparseLayout,
     SemiSparseLayout,
     TensorCoreTiledLayout,
-    UintxLayout,
     to_marlinqqq_quantized_intx,
+    UintxLayout,
 )
 from .utils import Layout, PlainLayout
